arm: Add a defconfig for arm a7 gnueabihf arch
This arch encompasses rpi2, orange pi plus/pc and pcDuino3 nano lite
boards in the cluster.

The defconfigs are generated from buildroot at around the 2015.08 release
with the build-overlay patch.

There is one defconfig for the toolchain, and one to use as a base for
other buildroot configs that use the toolchain (toolchain_client).