brp: Add buildroot-phase (brp) script
brp facilitates building buildroot projects in phases so that not
everything needs to be built at once. A final image can be created by
combining the output of previous phases.

This works best with my build-overlay and tar-overlay patches to
buildroot.

It's still not done, so include a TODO file as well.