brp: Introduce BRP_BOARD_DIR, move configs there
Move buildroot defconfigs from configs/ to a board-specific directory
and point at that with the exported variable BRP_BOARD_DIR.

A board directory should contain a toolchain symlink pointing at the
board directory for the toolchain "board" so that the toolchain_client
configuration can be found.

This change cleans up the configs/ directory and makes boards
self-contained.