brp: Add project, final build phase, small fixes.
Add <project> as the first command line argument so that we can add a
"core" directory at the same level as the project directory. brp is now
invoked from the top-level brp directory, not the project directory. The
core directory will contain common scripts, files, skeletons, etc for
all projects.

Add a "final" build phase that outputs a rootfs.tar image without a
phase name suffix, and that also invokes a build-image.sh script to
combine all parts and make the board-specific image.

Fix reading of defconfig - it cannot be sourced as it contains
make-style expansions $(FOO) which the shell treats as command
substitutions. Parse the defconfig with grep instead.

Do cleaner glob matching of rootfs images.