brp: Redo defconfig handling to remove config/ copy.
Using "make foo_defconfig" relies on the foo_defconfig file to be in
configs/. We don't keep them there, so we first need to copy it there,
and copy it out after a savedefconfig.

Instead of that, set BR2_DEFCONFIG to point straight to the defconfig
and just use "make defconfig". No copying needed.

Also invoke make once to do config targets instead of three times. Its
faster now.