brp: Move "final" image creation to post-image.sh
brp has a special case of a phase called "final" where it would run a
build-image.sh script in the board directory to do final assembly of an
image from the phase parts.

Instead of special-casing it, add an option to brp core
(BR2_BUILD_BOARD_IMAGE) which controls it, and move the logic to the
post-image.sh script.

This allows for the removal of the copy_images function and dependent
fucntions from brp, simplifying the top-level script.