brp/core: Add post-config script to alter config.
Add a post-config.sh script called by brp after a defconfig but before a
build allowing the .config file to be modified.

The post-config makes sure that the post-{build,fakeroot,image} scripts
are run and points at the "empty" skeleton if none is set. In future,
other changes may be made based on the core configuration - things that
cannot be done with Config.in selects or would otherwise require a
buildroot patch.