brp: Make output dirs for all targets.
Move the make_output_dirs function so we do it for all targets instead
of just certain ones. We were not doing it for unknown targets which is
when it is likely needed.

Also, only make output dirs if they don't exist, rather than relying on
mkdir -p. This means we do not generate unnecessary output.