brp: Put target first on command line.
Reorder command line arguments to put the target first. Target is
typically a verb and that makes sense first, with the object after.

Also handle the case when we are in a board/project directory building
for a different board/project. Previously we would assume a project was
not on the command line if we were in a project dir, similarly for
board. Now we only use the CWD defaults if they are not on the command
line.