kconfig: Add merge and demerge functions.
The merge and demerge functions are for managing fragments of defconfig
files.

For merging, all the fragments are merged in order with later
occurrences of variables overriding earlier ones.

For demerging, a set of fragments are subtracted from a complete
defconfig to produce just the config variables specific to the
defconfig. Whiteout entries are written (and processed on merge) so that
if a defconfig resets a var to default that has a non-default value in a
fragment, that reversal is not forgotten.