brp: Add defconfig fragment (de)merging.
Add A BRP_CONFIG_FRAGMENTS config setting that lists defconfig fragments
that should be merged into the current config to provide other settings.

When a config is editted (menuconfig, oldconfig), the fragments are
merged before the config and dermerged after the config to produce a
defconfig fragment containing just the settings specific to the phase
and omitting any settings made in any of the fragments.

The defconfig fragments are merged prior to performing a build.

The generated defconfig file is written to the phase output directory.

Note that the merge/demerge process is likely not foolproof. If you have
fragments with mutually exclusive options, and each with dependent
settings, they will not merge properly. The merge/demerge is not done
with reference to the Config.in files - it is a simple textual process.
You will need to make sure your fragments are compatible.