clusterpi: Add common sunxi image building and kconfig.
Define a pseudo-board "sunxi" to be the parent for the different
variations of boards in the pi cluster based on Allwinner SoCs.

Image building is done with the build-image.sh script which builds
images for NFS, sdcard and ramdisk root filesystems.

Common uboot config is provided, as is a linux defconfig for a kernel
common to all sunxi boards.