brp: Remove sdk host_dir special handling.
Remove the special handling of an existing BR2_HOST_DIR from the
sdk-specific build function and move it to the generic build function
for all cases where BR2_HOST_DIR is defined in the defconfig fragment.

Most defconfig fragments will not specify this and leave it default. In
those cases we don't worry about prompting about overwriting. But if it
is specified, it is assumed to be special and needs a prompt.