sunxi-fel: Add new package.
This is a subset of the sunxi-tools package in main buildroot, but main
buildroot has an old version of sunxi-fel that cannot boot a new (post
May 2017) u-boot image. There is not a released version of sunxi-tools
that can, so until there is, use a local package with a git commit hash
for the version.