sunxi: Update kernel configs to kernel 4.5.
This removes a common option (MEMCG_KMEM) as that has gone from the
kernel.

The sunxi changes incorporate the changes to the sunxi_defconfig in the
kernel. Note that the upstream defconfig added CONFIG_MFD_AXP20X_I2C and
CONFIG_MFD_AXP20X_RSB, but these symbols are not in the Kconfig. They
got added in 4.6, so that is when we'll add them.