sunxi: Add linux 4.6 defconfig.
This updates the defconfig with the changes in the sunxi_defconfig in
the kernel, adding the following settings (with their required
"parents"):

CONFIG_IR_SUNXI
CONFIG_SND_SUN4I_CODEC
CONFIG_USB_MUSB_SUNXI

Also added is:
CONFIG_MFD_AXP20X_{I2C,RSB}

These were added upstream to the defconfig in 4.5 but the kconfig did
not have them. The kconfig was added in 4.6, so we add to our defconfig
here.