sunxi: Add and use linux 4.8 configs.
The sunxi defconfig adds LEDS_TRIGGER_DISK - a new option that might be
useful for the disk nodes of the cluster.

Newly available, but not added is SND_SUN4I_I2S, as while it is sunxi, I
don't think any of my boards have a sound chip on the i2s bus. The chip
would need a specific driver too, so only enable this if necessary.

The default config brings in SUNXI_CCU and SUN8I_H3_CCU. I believe this
gets more use in later versions.