sunxi: Update kernel defconfig to 4.11, and use it.
This adds MFD_SUN4I_GPADC which appeared in the last kernel version but
we did not turn it on. It is now on so that THERMAL_OF can use it to get
CPU temps.

THERMAL_OF was added to the upstream sunxi_defconfig, but it is default
y, so we have not added it.

Also added is SND_SUN8I_CODEC, as a new sunxi config value. It should be
usable by the H3 socs on some of my boards.

By default, some sunxi PINCTL config options have changed - this just
seems to be keeping all of them set.

By default, SUN5I_CCU, SUN8I_V3S_CCU and SUN9I_A80_CCU are enabled.