uboot: Get NFS root parameters from dhcp.
Before booting NFS, run DHCP so we can get the nfs root parameters. This
allows a common boot image to be booted with different nfs root
directories (or servers) with the config driven by the DHCP server.

We need to turn off autoload so that the uboot dhcp command does not
proceed to try and load kernel images via TFTP. This makes the dhcp
command just to dhcp. autoload=0 has no other effect on uboot.