brp: Match all fragments, not just the first.
When searching for fragments, brp would stop at the first one found in
the parent chain and project "chain". Instead return all that match with
the most-specific fragments overriding the more general ones.

This means a parent can define a fragment called "linux" for instance,
with a board also defining "linux", and both fragments will be used.
Previously the parent would need to name the fragment something else for
the board to use it.