brp: Find project brp.conf correctly.
When searching for brp.conf we were looking in ${project}. This is a
relative path to the top of the BRP root, and will not work if we are in
a project or board directory already. This started failing when we
picked up the project and/or board from CWD and stopped requiring that
brp be run from the root.

Use ${BRP_PROJECT_DIR} instead and re-order when we load the config so
that it is set. Split out a couple of other var definitions as they
depend on the contents of the brp config.

While we're here, log a message when project/board/phase is missing
rather than just a usage message.