brp: Reset OPTIND for next getopts.
We run getopts twice - once for global options and once for command
options. getopts does not seem to reset OPTIND, so we need to reset it
ourselves, otherwise parsing command options in the presence of global
options will fail.

This was not noticed until now because there are no global options yet.
But when there are, it will become an issue.